clean up:   assertEquals(int, Integer) ambiguous.
Style.java: add findRenderer(Class)

PropertyChangeSupport: renamed to IPropertyChangeSupport

Ref.java: set<T> -> set<Object>

MethodSignature.java: incorrect hashcode.

Files.java: new classDataURL(), to support non-exist URLs (e.g. URL-prefix, URL-format, etc).

+++ bas/mod/Data Types/test/net/bodz/bas/types/IntRangeTest.java
@@ -10,11 +10,11 @@ public class IntRangeTest {
     public void test1() {
         IntRange range = new IntRange(10, 20);
         assertEquals(10, range.size());
-        assertEquals(10, range.ceiling(3));
-        assertEquals(19, range.floor(100));
+        assertEquals(10, (int) range.ceiling(3));
+        assertEquals(19, (int) range.floor(100));
         assertEquals(null, range.ceiling(20));
-        assertEquals(10, range.first());
-        assertEquals(19, range.last());
+        assertEquals(10, (int) range.first());
+        assertEquals(19, (int) range.last());
         range.add(9);
         assertEquals(11, range.size());
         for (int i = 19; i >= 9; i--)
